feat: implement Reddit Signal Scanner core features
- Backend: FastAPI + Celery + Reddit API integration
  * User authentication and authorization with JWT
  * Analysis task management with SSE streaming
  * Community pool management with Excel import/export
  * Warmup crawler with adaptive frequency control
  * 193 test cases covering core functionality
  * Type-safe implementation with mypy strict mode

- Frontend: React + TypeScript SPA
  * User authentication flow (register/login)
  * Analysis task creation and progress tracking
  * Real-time report display with SSE
  * Admin community pool management interface
  * Responsive design with Tailwind CSS
  * Full type safety with TypeScript strict mode

- Infrastructure:
  * Makefile for unified dev/test/deployment commands
  * Alembic database migrations (4 versions)
  * Docker Compose for test environment
  * Celery Beat for scheduled warmup tasks
  * Redis caching with adaptive hit rate monitoring

- Documentation:
  * Complete PRD (8 sections covering all features)
  * Spec-kit workflow specs (3 completed features)
  * Phase completion reports (Day 0-15)
  * Quality gates and acceptance criteria
  * MCP tools integration guide

Development Phases Completed:
- Phase 1-3: Warmup Period Implementation (Day 13-20)
  * Community pool with admin approval workflow
  * Adaptive crawler with cache hit rate optimization
  * Warmup period tracking and reporting

- Phase 4-5: Local Acceptance Testing
  * End-to-end testing with mock data
  * Frontend-backend integration verification
  * Quality gates validation

- Phase 6: Real Reddit API Integration
  * PRAW client integration
  * Rate limiting and error handling
  * Production-ready data collection

Technical Stack:
- Backend: Python 3.11, FastAPI 0.104, Celery 5.3, PostgreSQL, Redis
- Frontend: React 18, TypeScript 5, Vite 5, Tailwind CSS
- Testing: pytest (backend), vitest (frontend), Playwright (e2e)
- Quality: mypy strict, ESLint, Prettier, Black, isort

Ref: .specify/specs/001-day13-20-warmup-period/
Test Coverage: 193 backend tests
Code Quality: mypy strict mode (13 type warnings documented)

+++ backend/app/api/routes/admin.py
@@ -26,7 +26,7 @@ def _calculate_processing_seconds(task: Task) -> float | None:
     if task.started_at is None or task.completed_at is None:
         return None
     delta = task.completed_at - task.started_at
-    return round(delta.total_seconds(), 2)
+    return float(round(delta.total_seconds(), 2))
 
 
 def _collect_worker_count() -> int:
@@ -38,7 +38,7 @@ def _collect_worker_count() -> int:
     return len(active)
 
 
-@router.get("/dashboard/stats", summary="Admin dashboard aggregate metrics")
+@router.get("/dashboard/stats", summary="Admin dashboard aggregate metrics")  # type: ignore[misc]
 async def get_dashboard_stats(
     _payload: TokenPayload = Depends(require_admin),
     db: AsyncSession = Depends(get_session),
@@ -92,7 +92,7 @@ async def get_dashboard_stats(
     return _response(stats)
 
 
-@router.get("/tasks/recent", summary="Recent tasks overview")
+@router.get("/tasks/recent", summary="Recent tasks overview")  # type: ignore[misc]
 async def get_recent_tasks(
     _payload: TokenPayload = Depends(require_admin),
     limit: int = Query(50, ge=1, le=200),
@@ -123,7 +123,7 @@ async def get_recent_tasks(
     return _response({"items": items, "total": len(items)})
 
 
-@router.get("/users/active", summary="Active users ranked by recent tasks")
+@router.get("/users/active", summary="Active users ranked by recent tasks")  # type: ignore[misc]
 async def get_active_users(
     _payload: TokenPayload = Depends(require_admin),
     limit: int = Query(50, ge=1, le=200),
